feat: 支持main/slave模式

+++ controller/router.go
@@ -2,9 +2,11 @@ package controller
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/spf13/viper"
 )
 
 func InitRouter(r *gin.Engine) {
+	appMode := viper.GetString("App.Mode")
 	api := r.Group("/api")
 	{
 		// 服务管理
@@ -12,36 +14,40 @@ func InitRouter(r *gin.Engine) {
 		{
 			server.POST("/upload", Server.Upload)
 
-			server.POST("", Server.Create)
-			server.PUT("", Server.Update)
-			server.DELETE("", Server.Delete)
+			if appMode == "main" {
+				server.POST("", Server.Create)
+				server.PUT("", Server.Update)
+				server.DELETE("", Server.Delete)
 
-			// 服务节点管理
-			point := server.Group("/point")
-			{
-				point.GET("", ServerPoint.List)
-				point.POST("", ServerPoint.Create)
-				point.PUT("", ServerPoint.Update)
-				point.DELETE("", ServerPoint.Delete)
+				// 服务节点管理
+				point := server.Group("/point")
+				{
+					point.GET("", ServerPoint.List)
+					point.POST("", ServerPoint.Create)
+					point.PUT("", ServerPoint.Update)
+					point.DELETE("", ServerPoint.Delete)
+				}
 			}
 		}
 
-		// 文件夹管理
-		folder := api.Group("/folder")
-		{
-			folder.POST("", Folder.Create)
-			folder.GET("", Folder.List)
-			folder.PUT("", Folder.Update)
-			folder.DELETE("", Folder.Delete)
-		}
+		if appMode == "main" {
+			// 文件夹管理
+			folder := api.Group("/folder")
+			{
+				folder.POST("", Folder.Create)
+				folder.GET("", Folder.List)
+				folder.PUT("", Folder.Update)
+				folder.DELETE("", Folder.Delete)
+			}
 
-		// 节点管理
-		node := api.Group("/node")
-		{
-			node.POST("", Node.Create)
-			node.GET("", Node.List)
-			node.PUT("", Node.Update)
-			node.DELETE("", Node.Delete)
+			// 节点管理
+			node := api.Group("/node")
+			{
+				node.POST("", Node.Create)
+				node.GET("", Node.List)
+				node.PUT("", Node.Update)
+				node.DELETE("", Node.Delete)
+			}
 		}
 
 		docker := api.Group("/docker")
